+++ changelog.md
@@ -66,6 +66,9 @@ proc enumToString*(enums: openArray[enum]): string =
 ```
 
 - ``discard x`` is now illegal when `x` is a function symbol.
+- Implicit imports via ``--import: module`` in a config file are now restricted
+  to the main package.
+
 
 ### Library additions
 
@@ -91,6 +94,10 @@ proc enumToString*(enums: openArray[enum]): string =
   of two texts by line.
 
 - Added `os.relativePath`.
+- Added `parseopt.remainingArgs`.
+- Added `os.getCurrentCompilerExe` (implmented as `getAppFilename` at CT),
+  can be used to retrive the currently executing compiler.
+
 
 ### Library changes
 
@@ -110,6 +117,9 @@ proc enumToString*(enums: openArray[enum]): string =
 - The `parseopt` module now supports a new flag `allowWhitespaceAfterColon`
   (default value: true) that can be set to `false` for better Posix
   interoperability. (Bug #9619.)
+
+- `os.joinPath` and `os.normalizePath` handle edge cases like ``"a/b/../../.."``
+  differently.
 
 
 ### Language additions